Add RpcBlockResponse, refactor jsonrpc-primitives to have proper file structure
