+++ chain/jsonrpc-primitives/src/lib.rs
@@ -1,6 +1,6 @@
-pub mod blocks;
 pub mod errors;
 pub mod message;
 pub(crate) mod metrics;
 pub mod rpc;
+pub mod types;
 pub(crate) mod utils;
