+++ chain/jsonrpc-primitives/src/metrics.rs
@@ -5,7 +5,7 @@ lazy_static! {
     pub static ref RPC_UNREACHABLE_ERROR_COUNT: near_metrics::Result<IntCounterVec> =
         near_metrics::try_create_int_counter_vec(
             "near_rpc_unreachable_errors_total",
-            "Total count of Unexpected RPC errors returned, by method",
+            "Total count of Unreachable RPC errors returned, by method",
             &["method"]
         );
 }
