+++ chain/jsonrpc-primitives/src/types/mod.rs
@@ -1,0 +1,1 @@
+pub mod blocks;
